# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamPay API",
        "description": "High-integrity financial API for Santri Market Day ecosystem. Built with Laravel and secured with Sanctum.",
        "contact": {
            "name": "DreamPay Team",
            "email": "admin@dreampay.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.dreampay.id/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/merchant/scan-and-pay": {
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Merchant scans Santri QR Code to process payment",
                "operationId": "db67b6c34a89c033f589ba1ab5413ce3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_code",
                                    "amount"
                                ],
                                "properties": {
                                    "qr_code": {
                                        "type": "string",
                                        "example": "DP-ABC123XYZ"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 15000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Beli Nasi Goreng"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment successful"
                                        },
                                        "merchant_balance": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only merchants can perform this action"
                    },
                    "404": {
                        "description": "Santri not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/merchant/products": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "List merchant products",
                "operationId": "352bfc8eb33293bee70173611431de40",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Add a new product",
                "operationId": "2c96aa2b26d34a7a9eef3084fd964476",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Es Teh"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 3000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Es teh manis segar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/merchant/sales": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant sales history",
                "operationId": "cee85910271653119c9c90e195d75533",
                "responses": {
                    "200": {
                        "description": "Sales history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "merchant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nasi Goreng"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nasi goreng pedas sedap"
                    },
                    "price": {
                        "type": "number",
                        "example": 15000
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "title": "Transaction",
                "description": "Transaction model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "description": {
                        "type": "string",
                        "example": "Beli Nasi Goreng"
                    },
                    "type": {
                        "type": "string",
                        "example": "payment"
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "recipient_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}